+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './maps.js';
-export * from './matches.js';
-export * from './utils.js';
+export * from './maps';
+export * from './matches';
+export * from './types';